update from upstream
- fix some warnings for some weird versions of gcc
- cfgtool: remove 'rxtest' and 'rxraw' in favour of 'dump' and -n flag
- cfgtool: add 'status' command
- ff: add signal info to epoch
- ff: add trafo module (llh2xyz, xyz2llh, xyz2enu
- ff: refactor rxOpen() into rxInit() and rxOpen()

+++ ubloxcfg/ubloxcfg.c
@@ -873,7 +873,7 @@ static bool strToValUnsigned(const char *str, const UBLOXCFG_TYPE_t type, uint64
         return false;
     }
     const int len = strlen(str);
-    if ( (len < 1) || (isspace(str[0]) != 0) )
+    if ( (len < 1) || (isspace((int)str[0]) != 0) )
     {
         return false;
     }
@@ -948,7 +948,7 @@ static bool strToValSigned(const char *str, UBLOXCFG_TYPE_t type, int64_t *val)
         return false;
     }
     const int len = strlen(str);
-    if ( (len < 1) || (isspace(str[0]) != 0) )
+    if ( (len < 1) || (isspace((int)str[0]) != 0) )
     {
         return false;
     }
